Save action imported from module
