+++ assets/scripts/app.js
@@ -1,1 +1,3 @@
-// General JS
+/* eslint-disable no-unused-vars */
+import {CalculatorWidget} from './calculator/calculator.core';
+import {CalculatorResultLogger} from './calculator/calculator.logger';
